Make sure if there is an error in post processing it is recorded in the front end

+++ job_runner/job_runner_test/softlink/post_process.sh
@@ -13,7 +13,12 @@ do
    if [ "`expr $i % $TOTAL + 1`" == "$LSF_PM_TASKID" ]
    then
      python Convert1Dto2D.py $file >> out_$LSF_PM_TASKID.log
-     rm $file
+     if [ $? = 0 ]
+     then
+        rm $file
+     else
+       echo "[POST PROCESS ERROR] Post processing of $file failed"
+     fi
    fi
    i=`expr $i + 1`
 done
